Backend/requirements.txt
fastapi
uvicorn
python-dotenv
httpx
pydantic
firebase-admin
elevenlabs
